(resonance_finder) Completed third resonance finder
To find the resonance of two particles, where one particle is a tau
that decays into MET and another particle (jet, electron or muon),
the third decay finder function was implemented. It projects MET
along the visible decay products of the tau. From this projected
MET and the visible tau decay products the tau is reconstructed.
Between this reconstructed taus and the other particle list a
resonance is searched.