(n-1_plotts) Included Cut Class
To track the different properties of the cuts and the distributions of the variables
for each cut for each channel, the new Cuts Class was introduced which contains all
this information. Also the first function to create the N-1 distributions is created.
The first test distributions are created an written in the output file, but they are
not filled yet.