(EventNumbers) Included the furst part of the functionallity to have the event numbers saved
To keep track of the events that we want to track, a new map was included to keep
track of the accepted events there reconstructed resonance mas value for each channel
and the event information, like Runnumber, LumiSection and Eventnumber. A tree is created
to store this information. At the moment the filling of the tree is not working yet.
Also new Vertex histograms were included to cross check the pileup reweighting.